feat: make popup page dynamic so sandbox id can be included in the generated script tag
