+++ components/popup-page-dynamic.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+import dynamic from 'next/dynamic';
+
+export default dynamic(() => import('./popup-page'), { ssr: false });
